+++ router.go
@@ -7,23 +7,28 @@ import (
 	"github.com/gbrlsnchs/radix"
 )
 
+// Router is a router that implements a http.Handler.
 type Router struct {
 	prefix  string
 	methods map[string]*radix.Tree
 }
 
+// NewRouter creates and initializes a new Router.
 func NewRouter() *Router {
 	return &Router{methods: make(map[string]*radix.Tree)}
 }
 
+// Handle registers an http.Handler for a given method and path.
 func (rt *Router) Handle(m, p string, h http.Handler) {
 	rt.add(m, rt.prefix+resolvedPath(p), []interface{}{h})
 }
 
+// HandleFunc registers an http.HandlerFunc for a given method and path.
 func (rt *Router) HandleFunc(m, p string, hfunc http.HandlerFunc) {
 	rt.add(m, rt.prefix+resolvedPath(p), []interface{}{hfunc})
 }
 
+// HandleMiddlewares registers a stack of middlewares for a given method and path.
 func (rt *Router) HandleMiddlewares(m, p string, mids ...interface{}) {
 	midsToAdd := make([]interface{}, 0)
 
@@ -42,6 +47,10 @@ func (rt *Router) HandleMiddlewares(m, p string, mids ...interface{}) {
 	rt.add(m, rt.prefix+resolvedPath(p), midsToAdd)
 }
 
+// ServeHTTP uses a radix tree to find a given URL path from a request.
+//
+// It is able to pass route parameters using the request's context.
+// It can also break out of a middleware stack using the request's context cancelation.
 func (rt *Router) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	if rt.methods == nil || rt.methods[r.Method] == nil {
 		http.NotFound(w, r)
@@ -82,6 +91,7 @@ func (rt *Router) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	http.NotFound(w, r)
 }
 
+// Use registers a Subrouter to be used by the Router.
 func (rt *Router) Use(sub *Subrouter) {
 	for endp, mids := range sub.endps {
 		k := rt.prefix + endp
@@ -92,6 +102,8 @@ func (rt *Router) Use(sub *Subrouter) {
 	}
 }
 
+// WithPrefix sets a prefix for the Router, what makes
+// all registered handlers use the prefix set.
 func (rt *Router) WithPrefix(p string) *Router {
 	if p == "/" {
 		return rt
@@ -102,6 +114,7 @@ func (rt *Router) WithPrefix(p string) *Router {
 	return rt
 }
 
+// add adds middlewares to the radix tree.
 func (rt *Router) add(m, p string, mids []interface{}) {
 	if len(mids) == 0 || p == "" {
 		return