+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Makefile support.
+
+### Changed
+- Travis CI script.
+
+### Removed
+- Travis CI `goimports` script.
+
 ## [0.3.0] - 2017-11-11
 ### Added
 - Submux.
@@ -36,5 +46,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
+[Unreleased]: https://github.com/gbrlsnchs/httpmux/compare/v0.3.0...HEAD
 [0.3.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.1.0...v0.2.0