+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.4.0]
 ### Added
 - Makefile support.
 - `Router` and `Subrouter` structures.
@@ -55,6 +55,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - This package's source code, including examples and tests.
 - Go dep files.
 
-[Unreleased]: https://github.com/gbrlsnchs/httpmux/compare/v0.3.0...HEAD
+[0.4.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/gbrlsnchs/httpmux/compare/v0.1.0...v0.2.0